posts/requirements.txt
Django==3.0.8
psycopg2==2.8.5
psycopg2-binary==2.8.5
djangorestframework==3.11.0
dj-database-url==0.5.0
